refactor cli core validate preflight
